Move gherkin log steps to a separate file so that it can be shared
