+++ catalogue_graph/tests/adapters/transformers/folio/BDD/conftest.py
@@ -11,6 +11,7 @@ from models.pipeline.source.work import VisibleSourceWork
 
 # Allow * imports, pulling in individual step definitions is unwieldy
 # ruff: noqa: F403, F405
+from tests.gherkin_steps.logs import *
 from tests.gherkin_steps.marc import *
 from tests.gherkin_steps.work import *
